Create architecture skeleton
Created files (primarily interfaces) that demonstrate
how the controller works. In order to develop a feature,
you need to create an implementation for the request,
interactor, and response and tie them together in the
controller (please call the helper method
"executeRequest"). In addition, you need to modify the
authorizer to be able to handle your particular kind of
request. Perhaps this will be done with the visitor
pattern?